I. The following authorization policies have been added to the application - 1) 'MustBeMemberOfBoard': the user must be a member of the board for the following  API: - in BoardsController: GetBordById ([get]:api/Boards/boardId), AddColumnToBoard ([post]:api/Boards/boardId/columns); - in ColumnsController: Update ([put]:api/Columns/columnId); - in IssuesController: MoveIssue ([patch]:api/Issues/issueId/move); 2) 'MustBeBoardAdminAddUserToBoard': - in BoardsController: AddUserToBoardCommand ([post]:api/Boards/boardId/members), RemoveUserFromBoard (delete]:api/Boards/boardId/members), DeleteBordById (delete]:api/Boards/boardId); - in ColumnsController: Delete ([delete]:api/Columns/columnId), AddIssueToColumn ([post]:api/Columns/columnId/issues) 3) 'MustBeIssueCreatorOrAssigneeOrAdmin': - in IssuesController: Delete ([delete]:api/Issues/issueId); 4) MustBeIssueAssigneeOrAdminOrIssueAssigneeIsNull: - in IssuesController: Update ([put]:api/Issues/issueId). II. Added the following classes: - ValidateIdAttribute for validation of input params in the controller actions; - IEntityRepository - interface with the general methods of the entity repositories; - JsonSerializerOptionsExtensions - added the exstension method for getting the default JsonSerializerOptions; - StatusCodeExtensions - adding the exstension method for getting StatusCode; - AuthorizationMiddlewareResultHandler - for processing the Authorization Failures. III. Improved the following classes: - GlobalExceptionHandler - used the exstension GetTitleForStatus() from the StatusCodeExtensions class; - Startup - added policies and AuthorizationHandlers; - BoardsController, ColumnController, IssuesController - fixed actions and added attributes with policies; - 01-Tables.sql, 02-SprocsForCrud.sql and 03-SprocsForBusinessLogic.sql - fixed storage procedures; - fixed SqlStatements class and repositories and interfaces for them; - fixed dtos, queries, commands and handlers for the Mediator pattern.

+++ Board.BusinessLogic/Features/ForIssue/Queries/GetIssueByIdQuery.cs
@@ -3,4 +3,4 @@ using MediatR;
 
 namespace Board.BusinessLogic.Features.ForIssue.Queries;
 
-public record GetIssueByIdQuery(int Id) : IRequest<IssueResponseDto>;
+public record GetIssueByIdQuery(long Id) : IRequest<IssueResponseDto>;
